+++ Source/Applications/Desktop/VodovozViewModels/ReportsParameters/Logistic/DeliveriesLateReportViewModel.OrderSelectMode.cs
@@ -1,0 +1,12 @@
+﻿namespace Vodovoz.ViewModels.ReportsParameters.Logistic
+{
+	public partial class DeliveriesLateReportViewModel
+	{
+		private enum OrderSelectMode
+		{
+			All,
+			DeliveryInAnHour,
+			WithoutDeliveryInAnHour
+		}
+	}
+}
